examples/hyphen-chat/requirements.txt
# Requirements for hyphen-chat
pycryptodome>=3.20.0
asynctkinter>=0.3.0
